[Concept Lab] 支持 4 GiB 单文件 (#95)
* fs: define 64-bit triple-indirect inode layout

* fs: widen inode size and open-file offset

* fs: use 64-bit inode read and write offsets

* fs: size the log for atomic 4 GiB unlink

* fs: add multi-block crash-safe log headers

* fs: propagate 64-bit offsets and short writes

* fs: add triple-indirect 64-bit block mapping

* build: add sparse 4 GiB regression image

* test: keep Lab9 bigfile bounded after MAXFILE growth

* test: add explicit 4 GiB host runner

* test: validate isolated large-file suite wiring

* user: print 64-bit file sizes in ls

* user: add correct unsigned 64-bit printf formatting

* test: cover 4 GiB write read stat full unlink and reuse

* test: register isolated 4 GiB guest regression

* build: create large sparse file-system images

* docs: document explicit 4 GiB regression workflow

* test: minimize large-file registry diff

* build: ignore large-image mkfs binary

* mkfs: encode 64-bit inode sizes in normal images

+++ kernel/defs.h
@@ -54,9 +54,9 @@ void            iupdate(struct inode*);
 int             namecmp(const char*, const char*);
 struct inode*   namei(char*);
 struct inode*   nameiparent(char*, char*);
-int             readi(struct inode*, int, uint64, uint, uint);
+int             readi(struct inode*, int, uint64, uint64, uint);
 void            stati(struct inode*, struct stat*);
-int             writei(struct inode*, int, uint64, uint, uint);
+int             writei(struct inode*, int, uint64, uint64, uint);
 void            itrunc(struct inode*);
 
 // ramdisk.c
@@ -189,7 +189,7 @@ void            vmwalk(pagetable_t pagetable, int depth);
 void            vmprint(pagetable_t pagetable);
 void            kvmmapkern(pagetable_t pagetable, uint64 va, uint64 pa, uint64 sz, int perm);
 pagetable_t     kvmcreate();
-void            kvmfree(pagetable_t kpagetale) ;
+void            kvmfree(pagetable_t kpagetale);
 void            u2kvmcopy(pagetable_t, pagetable_t, uint64, uint64);
 void            u2kvmunmap(pagetable_t, uint64, uint64);
 int             uvmlazyalloc(struct proc*, uint64);
@@ -207,6 +207,7 @@ void            virtio_disk_intr(void);
 
 uint64          free_proc(void);
 uint64          free_mem(void);
+
 // vma.c
 void            vma_init(void);
 struct VMA*     vma_alloc(void);
